Add files via upload
This is an introductory paper I wrote about BPM/process mining, during the early periods of my education at Avans Hogeschool of Applied Sciences. It clearly shows I had no clue on how to write a paper (yet), but it also illustrates my enthusiasm for the subject. Read with a grain of salt! :D